adding in some of the basic UI and functionality for increasing weapon damage as well as utilizing weapon damages on takedamage

+++ Assets/Scripts/HealthBar.cs
@@ -66,9 +66,9 @@ public class HealthBar : MonoBehaviour
 
     }
 
-    public int TakeDamage(int damageTaken)
+    public int TakeDamage(float damageTaken)
     {
-        currentHealth -= damageTaken;
+        currentHealth -= (int)damageTaken;
         AnimationController();
         RenderCheck();
         return currentHealth;
